Change evenSum to long long for larger values
Change the type of evenSum from int to long long to handle larger sums.

+++ Special_Sum_of_a_Sequence.cpp
@@ -6,7 +6,7 @@ int main(){
     int a, d, b;
     cin>>a>>d>>b;
 
-    int evenSum=0;
+    long long evenSum=0;
     cout<<"Sequence: ";
     for(int i=a; i<=b; i+=d){
